ParseFileNameError to dedicated file
