+++ src/parser.rs
@@ -1,38 +1,7 @@
+use super::parse_file_name_error::ParseFileNameError;
+
 use super::file_entry::FileEntry;
-use colored::Colorize;
 use std::collections::HashMap;
-use std::error::Error;
-use std::fmt;
-
-#[derive(Debug, Clone)]
-pub struct ParseFileNameError {
-    file_name: String,
-}
-
-impl fmt::Display for ParseFileNameError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "{}: '{}' doesn't match the pattern! (<TICKET_NUMBER>_<SECTION>_...)",
-            "ERROR".red(),
-            self.file_name
-        )
-    }
-}
-
-impl Error for ParseFileNameError {}
-
-impl From<String> for ParseFileNameError {
-    fn from(file_name: String) -> Self {
-        ParseFileNameError { file_name }
-    }
-}
-
-impl ParseFileNameError {
-    fn new(file_name: String) -> ParseFileNameError {
-        ParseFileNameError { file_name }
-    }
-}
 
 pub fn parse_file_name(name: String) -> Result<(String, String), ParseFileNameError> {
     let splitted: Vec<&str> = name.split('_').collect();
